matrix: change to use at e-mail
Gmail and I don't know who else make things ugly when you use
relative+absolute. This is probably because of safety, with absolute
someone could override part of your e-mail screen.

So now the parts will have all the same size and will be used as
background of divs one inside the other.